+++ src/core/localize/localize.models.ts
@@ -68,6 +68,8 @@ export enum Localization {
   cancel = 'cancel',
   clear = 'clear',
   confirm = 'confirm',
+  on = 'on',
+  off = 'off',
   change_language = 'change_language',
   change_language_description = 'change_language_description',
   language = 'language',