+++ src/more/more.navigation.tsx
@@ -8,6 +8,8 @@ import { Author } from './components/author/author'
 import { DeveloperNavigation } from '../developer/developer.navigation'
 import { User } from '../user/user'
 import { Language } from './components/language/language'
+import { Theme } from './components/theme/theme'
+import { HapticFeedback } from './components/haptic-feedback/haptic-feedback'
 
 const Stack = createNativeStackNavigator()
 
@@ -15,6 +17,8 @@ export const MoreNavigation = () => (
   <Stack.Navigator screenOptions={DEFAULT_SCREEN_OPTIONS}>
     <Stack.Screen name={SCREEN.MORE_MAIN} component={More} />
     <Stack.Screen name={SCREEN.MORE_LANGUAGE} component={Language} />
+    <Stack.Screen name={SCREEN.MORE_THEME} component={Theme} />
+    <Stack.Screen name={SCREEN.MORE_HAPTIC} component={HapticFeedback} />
     <Stack.Screen name={SCREEN.MORE_MANIA} component={Mania} />
     <Stack.Screen name={SCREEN.MORE_HELP} component={HelpNavigation} />
     <Stack.Screen name={SCREEN.MORE_AUTHOR} component={Author} />